+++ src/textual/widget.py
@@ -2571,13 +2571,13 @@ class Widget(DOMNode):
             return Text(renderable)
         return renderable
 
-    async def action(self, action: str) -> None:
+    async def run_action(self, action: str) -> None:
         """Perform a given action, with this widget as the default namespace.
 
         Args:
             action: Action encoded as a string.
         """
-        await self.app.action(action, self)
+        await self.app.run_action(action, self)
 
     def post_message(self, message: Message) -> bool:
         """Post a message to this widget.
@@ -2602,6 +2602,9 @@ class Widget(DOMNode):
         Args:
             event: Idle event.
         """
+        self._check_refresh()
+
+    def _check_refresh(self) -> None:
         if self._parent is not None and not self._closing:
             try:
                 screen = self.screen